Eco mode page

+++ src/pages/EcoMode.tsx
@@ -1,38 +1,43 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import styled from 'styled-components';
-import TokenListForm from '../components/TokenListForm';
+import EcoModeForm from '../components/EcoModeForm';
 import CustomConfirmationModal from '../components/CustomConfirmationModal';
 import deleteIcon from '../assets/delete.png';
 import { toast } from 'react-toastify';
 import { ContainerTable } from '../utils/styles';
+import Network from "../utils/Network";
 import moment from 'moment';
 import { refresh } from '../utils/web3-utils';
 
-interface TokenList {
-  isActive: boolean;
-  lastUpdate: string;
-  name: string;
-  updatedAt: string;
-  url: string;
+interface EcoMode {
+  active: boolean;
+  safeGuardPeriodPct: number;
+  averageSuccessSpeedSmoothFactor: number;
+  maximumExecutionsPerPeriod: number;
   _id: string;
+  updatedAt: string;
+  smartVault: {
+   address: string;
+   chainId: number;
+ };
 }
 
 const URL = process.env.REACT_APP_SERVER_BASE_URL;
 
-const TokenLists: React.FC = () => {
+const EcoModes: React.FC = () => {
   const [
-    tokenRegistryData,
-    setTokenLists,
-  ] = useState<TokenList[] | null>(null);
+    ecoModesData,
+    setEcoModes,
+  ] = useState<EcoMode[] | null>(null);
   const [customModalOpen, setCustomModalOpen] = useState(false);
-  const [deleteParams, setDeleteParams] = useState<string>('');
+  const [deleteParams, setDeleteParams] = useState<EcoMode | null>(null);
 
-  const fetchTokenList = async () => {
+  const fetchEcoModes = async () => {
     try {
       const token = localStorage.getItem('token');
-      const response = await axios.get<TokenList[]>(
-        `${URL}/token-registry/lists`,
+      const response = await axios.get<EcoMode[]>(
+        `${URL}/relayer-executor/eco-modes`,
         {
           headers: {
             'Access-Control-Allow-Origin': '*',
@@ -42,12 +47,12 @@ const TokenLists: React.FC = () => {
         }
       );
 
-      setTokenLists(response.data);
+      setEcoModes(response.data);
     } catch (error: any) {
       if (error.response?.status === 401) {
         try {
           await refresh();
-          await fetchTokenList();
+          await fetchEcoModes();
         } catch (refreshError) {
           console.error(`Error: Unable to refresh token. Please log in again.`);
         }
@@ -57,40 +62,41 @@ const TokenLists: React.FC = () => {
   };
 
   useEffect(() => {
-    fetchTokenList();
+    fetchEcoModes();
   }, []);
 
-  const handleDeleteClick = (id: string) => {
-    setDeleteParams(id);
+  const handleDeleteClick = (item: EcoMode) => {
+    setDeleteParams(item);
     setCustomModalOpen(true);
   };
 
   const handleConfirmDelete = async () => {
-    const id = deleteParams;
     const token = localStorage.getItem('token');
-    const url = `${URL}/token-registry/lists/${id}`;
+    if(deleteParams) {
+      const url = `${URL}/relayer-executor/eco-modes/${deleteParams.smartVault.chainId}/${deleteParams.smartVault.address}`;
 
-    try {
-      await axios.delete(url, {
-        headers: {
-          'Access-Control-Allow-Origin': '*',
-          'Content-Type': 'application/json',
-          'x-auth-token': `${token}`,
-        },
-      });
-      fetchTokenList();
-      toast.success('Token list successfully deleted');
+      try {
+        await axios.delete(url, {
+          headers: {
+            'Access-Control-Allow-Origin': '*',
+            'Content-Type': 'application/json',
+            'x-auth-token': `${token}`,
+          },
+        });
+        fetchEcoModes();
+        toast.success('Eco mode successfully deleted');
 
-    } catch (error: any) {
-      if (error.response?.status === 401) {
-        try {
-          await refresh();
-          await handleConfirmDelete();
-        } catch (refreshError) {
-          console.error(`Error: Unable to refresh token. Please log in again.`);
+      } catch (error: any) {
+        if (error.response?.status === 401) {
+          try {
+            await refresh();
+            await handleConfirmDelete();
+          } catch (refreshError) {
+            console.error(`Error: Unable to refresh token. Please log in again.`);
+          }
         }
+        console.error('There was an error deleting this eco mode item:', error);
       }
-      console.error('There was an error deleting the token list item:', error);
     }
 
     setCustomModalOpen(false);
@@ -102,37 +108,41 @@ const TokenLists: React.FC = () => {
 
 
   return (
-    <TokenListSection>
-      <TokenListForm onSuccess={fetchTokenList} />
-      {tokenRegistryData ? (
+    <EcoModeSection>
+      <EcoModeForm onSuccess={fetchEcoModes} />
+      {ecoModesData ? (
         <>
           <ContainerTable>
             <thead>
               <tr>
-                <th>Name</th>
-                <th>URL</th>
+                <th>Smart Vault</th>
+                <th>Network</th>
                 <th>Active</th>
+                <th>Success Speed Smooth Factor</th>
+                <th>Safe Guard Period Pct</th>
                 <th>Last Update</th>
                 <th></th>
               </tr>
             </thead>
             <tbody>
-              {tokenRegistryData.map((item, index) => (
+              {ecoModesData.map((item, index) => (
                 <tr key={index}>
-                  <td>{item.name}</td>
+                  <td>{item.smartVault.address}</td>
+                  <td><Network network={item.smartVault.chainId} width={1200} /></td>
                   <td>
-                    {item.url}
+                    {item.active ? '🟢' : '🔴'}
                   </td>
                   <td>
-                    {item.isActive ? '🟢' : '🔴'}
+                    {item.averageSuccessSpeedSmoothFactor}
                   </td>
+                  <td>{item.safeGuardPeriodPct}</td>
                   <td>
-                    {moment(item.lastUpdate).format('MMMM DD, YYYY [at] HH:mm:ss')}
+                    {moment(item.updatedAt).format('MMMM DD, YYYY [at] HH:mm:ss')}
                   </td>
                   <td>
                   <img
                     onClick={() =>
-                      handleDeleteClick(item._id)
+                      handleDeleteClick(item)
                     }
                     src={deleteIcon}
                     alt="Delete"
@@ -153,11 +163,11 @@ const TokenLists: React.FC = () => {
       ) : (
         <p>Loading...</p>
       )}
-    </TokenListSection>
+    </EcoModeSection>
   );
 };
 
-const TokenListSection = styled.div`
+const EcoModeSection = styled.div`
   margin: 0px auto;
   display: flex;
   flex-direction: column;
@@ -166,4 +176,4 @@ const TokenListSection = styled.div`
   max-width: 90%;
 `;
 
-export default TokenLists;
+export default EcoModes;
